Split up fixup.cpp a bit
